Remove fetch X movie details from api service class
Delete ApiService class
