+++ shared/src/commonMain/kotlin/com/vickikbt/shared/data/datasources/MovieDetailsRepositoryImpl.kt
@@ -1,19 +1,22 @@
 package com.vickikbt.shared.data.datasources
 
-import com.vickikbt.shared.data.cache.sqldelight.daos.MovieDao
 import com.vickikbt.shared.data.mappers.toDomain
-import com.vickikbt.shared.data.mappers.toEntity
-import com.vickikbt.shared.data.network.ApiService
+import com.vickikbt.shared.data.network.models.CastDto
+import com.vickikbt.shared.data.network.models.MovieDetailsDto
+import com.vickikbt.shared.data.network.models.MovieResultsDto
 import com.vickikbt.shared.data.network.utils.safeApiCall
 import com.vickikbt.shared.domain.models.Cast
 import com.vickikbt.shared.domain.models.Movie
 import com.vickikbt.shared.domain.models.MovieDetails
 import com.vickikbt.shared.domain.repositories.MovieDetailsRepository
+import io.ktor.client.HttpClient
+import io.ktor.client.call.body
+import io.ktor.client.request.get
+import io.ktor.client.request.parameter
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.map
 
 class MovieDetailsRepositoryImpl constructor(
-    private val apiService: ApiService,
+    private val httpClient: HttpClient,
     // private val movieDao: MovieDao
 ) : MovieDetailsRepository {
 
@@ -28,20 +31,29 @@ class MovieDetailsRepositoryImpl constructor(
         return fetchMovieDetails(movieId = movieId)
     }*/
 
-    override suspend fun fetchMovieDetails(movieId: Int): Flow<Result<MovieDetails>> = safeApiCall {
-        val networkResponse = apiService.fetchMovieDetails(movieId = movieId)
-
-        return@safeApiCall networkResponse.toDomain()
+    override suspend fun fetchMovieDetails(movieId: Int): Flow<Result<MovieDetails>> {
+        return safeApiCall {
+            val response = httpClient.get(urlString = "movie/$movieId").body<MovieDetailsDto>()
+            response.toDomain()
+        }
     }
 
-    override suspend fun fetchMovieCast(movieId: Int): Flow<Result<Cast>> = safeApiCall {
-        val movieCastNetworkResponse = apiService.fetchMovieCast(movieId)
+    override suspend fun fetchMovieCast(movieId: Int): Flow<Result<Cast>> {
+        return safeApiCall {
+            val response = httpClient.get(urlString = "movie/$movieId/credits").body<CastDto>()
 
-        return@safeApiCall movieCastNetworkResponse.toDomain()
+            response.toDomain()
+        }
     }
 
-    override suspend fun fetchSimilarMovies(movieId: Int): Flow<Result<List<Movie>?>> = safeApiCall {
-        val similarMoviesDto = apiService.fetchSimilarMovies(movieId).movies?.map { it.toEntity() }
-        return@safeApiCall similarMoviesDto?.map { it.toDomain() }
+    override suspend fun fetchSimilarMovies(movieId: Int, page: Int): Flow<Result<List<Movie>?>> {
+        return safeApiCall {
+            val response = httpClient.get(urlString = "movie/$movieId/similar") {
+                parameter("page", page)
+            }.body<MovieResultsDto>()
+
+            response.movies?.map { it.toDomain() }
+        }
     }
+
 }
