+++ shared/src/commonMain/kotlin/com/vickikbt/shared/domain/repositories/MovieDetailsRepository.kt
@@ -3,6 +3,7 @@ package com.vickikbt.shared.domain.repositories
 import com.vickikbt.shared.domain.models.Cast
 import com.vickikbt.shared.domain.models.Movie
 import com.vickikbt.shared.domain.models.MovieDetails
+import com.vickikbt.shared.domain.utils.Constants.STARTING_PAGE_INDEX
 import kotlinx.coroutines.flow.Flow
 
 interface MovieDetailsRepository {
@@ -25,5 +26,8 @@ interface MovieDetailsRepository {
     suspend fun fetchMovieCast(movieId: Int): Flow<Result<Cast>>
 
     /** Fetches similar movies from network source*/
-    suspend fun fetchSimilarMovies(movieId: Int): Flow<Result<List<Movie>?>>
+    suspend fun fetchSimilarMovies(
+        movieId: Int,
+        page: Int = STARTING_PAGE_INDEX
+    ): Flow<Result<List<Movie>?>>
 }
